refactor: restructure agent modules + documentation foundations
Phase 0 — Documentation + Instructions:
- Create testing.instructions.md with modern testing best practices
- Fix stale python.instructions.md (wrong architecture, wrong file names)
- Fix stale docs.instructions.md (P1-P8 → P1-P12, R1-R20 → R1-R27)
- Add ADR guidance to copilot-instructions.md and docs.instructions.md
- Write ADR-005: CI access pattern (Tailscale ephemeral + scoped curl)
- Write ADR-006: Dokploy GitOps (auto-deploy with UI env vars)

Phase 1 — Agent code refactoring:
- Extract github.py — merge github_app.py auth with review.py API helpers
- Rename copilot_cli.py → copilot.py
- Rename worktree.py → git.py
- Slim review.py to pure orchestration (~95 lines, was ~287)
- Update and expand tests (23 passing, +2 new bot_login tests)
- Delete github_app.py (code merged into github.py)
