+++ .github/instructions/python.instructions.md
@@ -11,19 +11,21 @@ Style is enforced by ruff and ty — see `stacks/agents/app/pyproject.toml` for 
 The agent service (`stacks/agents/app/`) is a FastAPI app that:
 
 1. Receives requests (e.g., `/review`) with minimal input (repo + PR number)
-2. Calls external APIs (GitHub, Copilot) using local credentials (`gh auth token`)
-3. Returns structured JSON responses — the agent never writes to GitHub directly
-4. GitHub Actions workflows handle posting reviews, comments, etc.
+2. Runs Copilot CLI in headless mode to perform the work (review, implement)
+3. Copilot CLI interacts with GitHub directly — posting reviews, reading files, etc.
 
 ### Key files
 
 - `main.py` — FastAPI endpoints, request/response models
-- `review.py` — PR review logic, system prompt, structured output parsing
-- `copilot.py` — Copilot API client, returns `ChatResult` dataclass
-- `tests/test_main.py` — unit tests (mock external calls)
+- `review.py` — PR review orchestrator (deep module, narrow interface)
+- `copilot.py` — Copilot CLI subprocess wrapper, returns `CLIResult` with parsed stats
+- `git.py` — Git operations: bare clone, worktrees, branches, commit and push
+- `github.py` — GitHub API: App auth (JWT → installation token), REST, GraphQL helpers
+- `tests/` — unit tests, one per module (mock external calls at boundaries)
 
 ## Testing
 
-- Mock external calls via `@patch("main.review_pr", new_callable=AsyncMock)`
+- Mock external calls at boundaries: `@patch("main.review_pr", new_callable=AsyncMock)`
 - Use `TestClient(app)` from FastAPI for endpoint tests
 - Run from the agent directory: `cd stacks/agents/app && uv run pytest`
+- See `.github/instructions/testing.instructions.md` for full testing guidelines
